+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.3.0] - 2024-12-24 (minor)
 
 - Fix YouTube embeds to Decidim v0.28 format in different places. Only YouTube is supported right now.
+- Add task to add new content blocks in participatory processes
 
 ## [0.2.1] - 2024-09-25 (patch)